Restarting major refactoring of client
